requirements.txt
azure-cognitiveservices-speech==1.42.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-storage-blob==12.25.0
openai==1.67.0
pydantic==2.10.6
requests==2.32.3
